The list of things I did today
So I was able to put all six boxes into our website (I think that is the only thing we should do for now). The only thing I am stuck on is centering the yellow box and trying not to have the orange bar overlapping the top left blue box. At least we are almost finish, just need to fix some things up and call it a day.